pyproject.toml
[project]
name = "agent-marketplace"
version = "0.1.0"
description = "Multi-agent marketplace with ANS, escrow, and satoshi micropayments"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
]
